+++ NEWS.md
@@ -1,3 +1,11 @@
+# eplusr 0.8.1.1
+
+To be released as 0.8.2
+
+* `$get_value()` in `IdfObject` class new has a new argument `simplify`. If
+  `TRUE`, a character vecotr will be returned instead of a named list. Default
+  is `FALSE`.
+
 # eplusr 0.8.1
 
 ## Major changes